Made levels load more dynamically, updated combat info
Renamed World to Level

+++ src/BK9K/BK9K.Game/Types/ClassTypes.cs
@@ -2,10 +2,10 @@
 {
     public class ClassTypes
     {
-        public static readonly byte Unknown = 0;
+        public static readonly int Unknown = 0;
 
-        public static readonly byte Fighter = 1;
-        public static readonly byte Mage = 2;
-        public static readonly byte Rogue = 3;
+        public static readonly int Fighter = 1;
+        public static readonly int Mage = 2;
+        public static readonly int Rogue = 3;
     }
 }
